+++ resources/[dg]/dg-misc/typescript/src/server/controllers/exports.ts
@@ -1,0 +1,3 @@
+import { getChunkForPos } from '../../shared/helpers/grid';
+
+global.exports('getChunkForPos', getChunkForPos);