+++ resources/[dg]/ts-shared/typescript/src/client/classes/util.ts
@@ -46,6 +46,9 @@ class Util extends UtilShared {
   }
 
   loadModel = (model: string | number) => {
+    if (typeof model === 'string') {
+      model = GetHashKey(model)
+    }
     RequestModel(model);
     return this.awaitCondition(() => HasModelLoaded(model));
   };
@@ -488,19 +491,19 @@ export class Animations {
   };
 }
 
-class StaticObjects {
+class SyncedObjects {
   private objectsToRemove: Set<string>;
 
   constructor() {
     this.objectsToRemove = new Set();
     on('onResourceStop', (res: string) => {
       if (res !== GetCurrentResourceName()) return;
-      global.exports['dg-misc'].removeStaticObject([...this.objectsToRemove]);
+      global.exports['dg-misc'].removeObject([...this.objectsToRemove]);
     });
   }
 
-  public add = async (objectData: StaticObjects.CreateData | StaticObjects.CreateData[]): Promise<string[]> => {
-    const ids: string[] = await global.exports['dg-misc'].addStaticObject(objectData);
+  public add = (objectData: Objects.CreateData | Objects.CreateData[]): string[] => {
+    const ids: string[] = global.exports['dg-misc'].addLocalObject(objectData);
 
     ids.forEach(id => {
       this.objectsToRemove.add(id);
@@ -515,7 +518,7 @@ class StaticObjects {
       this.objectsToRemove.delete(objId);
     }
 
-    global.exports['dg-misc'].removeStaticObject(objId);
+    global.exports['dg-misc'].removeObject(objId);
   };
 
   public getEntityForObjectId = (objId: string): number | undefined => {
@@ -529,5 +532,5 @@ export default {
   PropAttach: new PropAttach(),
   Particle: new Particle(),
   Animations: new Animations(),
-  StaticObjects: new StaticObjects(),
+  SyncedObjects: new SyncedObjects(),
 };